all input parsing now happens through separate thread node.InputReader() watcher now runs, will alert us of filesystem changes still more work to do here

+++ node/send.go
@@ -53,7 +53,7 @@ func (n *Node) SendFile(path string) error {
 
 		push := &commands.Push{
 			Path:       path,
-			Length:     int64(len),
+			Length:     len,
 			ModifiedAt: info.ModTime().Unix(),
 			Mode:       mode.Perm(),
 			More:       more,
